Replace old looper functions with dope new ones
