+++ looper/loop.go
@@ -5,51 +5,27 @@
 package looper
 
 import (
-	"github.com/emer/emergent/etime"
-	"github.com/goki/ki/indent"
+	"github.com/emer/emergent/envlp"
 )
 
-//THIS IS EXPERIMENTAL and UNFINISHED AND LIKELY WILL CHANGE
-type LoopPart string
+type LoopStructure struct {
+	OnStart orderedMapFuncs
+	// Either Main or the inner loop occurs between OnStart and OnEnd
+	Main   orderedMapFuncs
+	OnEnd  orderedMapFuncs
+	IsDone map[string]func() bool `desc:"If true, end loop. Maintained as an unordered map because they should not have side effects."`
 
-const (
-	End  LoopPart = "END"
-	Stop LoopPart = "STOP"
-	Main LoopPart = "MAIN"
-	Init LoopPart = "Cycle0"
-)
+	Phases []Phase `desc:"Only use Phases at the Theta Cycle timescale (200ms)."`
+	// TODO Add an axon.time here but move it to etimes
 
-func (loop LoopPart) String(part etime.Times) string {
-	return string(part) + ":" + string(loop)
+	Counter *envlp.Ctr `desc:"Tracks time within the loop. Also tracks the maximum."`
 }
 
-// Loop represents one level of looping, with arbitrary functions
-// called at 3 different points in the loop, corresponding to a
-// do..while loop logic, with no initialization, which is necessary
-// to ensure reentrant steppability.  In Go, the logic looks like this:
-//
-// for {
-//    for { <subloops here> } // drills down levels for each subloop
-//    Main()                  // Main is called after subloops -- increment counters!
-//    if Stop() {
-//        break
-//    }
-// }
-// End()                      // Reset counters here so next pass starts over
-//
-type Loop struct {
-	Stack *Stack         `desc:"stack that owns this loop"`
-	Scope etime.ScopeKey `desc:"scope level of this loop"`
-	Main  Funcs          `desc:"main functions to call inside each iteration, after looping at lower level for non-terminal levels -- any counters should be incremented here -- if there is an Env set for the Stack, then any counter in the Env at the corresponding Scope will automatically be incremented via Env:Incr or Env:Step functions added automatically"`
-	Stop  BoolFuncs      `desc:"functions that cause the loop to stop -- if any return true, it stops"`
-	End   Funcs          `desc:"functions to run at the end of the loop, after it has stopped.  counters etc should be reset here, so next iteration starts over afresh.  the Init function calls these to initialize before running."`
-}
-
-func NewLoop(sc etime.ScopeKey, st *Stack) *Loop {
-	return &Loop{Scope: sc, Stack: st}
-}
-
-// StageString returns a string for given stage of loop, indented to level
-func (lp *Loop) StageString(stage string, level int) string {
-	return indent.Spaces(level, indentSize) + string(lp.Scope) + ": " + stage
+func (loops *LoopStructure) AddPhases(phases ...Phase) {
+	for _, phase := range phases {
+		loops.Phases = append(loops.Phases, phase)
+		phase.OnMillisecondEnd = orderedMapFuncs{}
+		phase.PhaseStart = orderedMapFuncs{}
+		phase.PhaseEnd = orderedMapFuncs{}
+	}
 }
